+++ novelservice/src/main/java/com/iucyh/novelservice/episode/repository/custom/condition/EpisodePagingCondition.java
@@ -1,4 +1,4 @@
-package com.iucyh.novelservice.episode.repository.query.condition;
+package com.iucyh.novelservice.episode.repository.custom.condition;
 
 import com.iucyh.novelservice.episode.enumtype.EpisodeSortType;
 import org.springframework.data.domain.Pageable;
